На прорверку задание 4.1

+++ src/main/java/ru/hogwarts/school/service/StudentService.java
@@ -31,7 +31,7 @@ public class StudentService {
         studentRepository.deleteById(studentId);
     }
 
-    public List <Student> findByAge (Long age){
+    public List <Student> findByAge (Integer age){
         return studentRepository.findByAge(age);
     }
